Crate clmul

CarryLess MULtiplication (clmul) based on the crate: https://github.com/RustCrypto/universal-hashes/tree/master/polyval

Only those comments from the original file are preserved which are relevant to carryless multiplication.

Minimum Supported Rust Version

Rust 1.56 or higher.

Supported backends

This crate provides multiple backends including a portable pure Rust backend as well as ones based on CPU intrinsics.

“soft” portable backend

As a baseline implementation, this crate provides a constant-time pure Rust implementation based on [BearSSL], which is a straightforward and compact implementation which uses a clever but simple technique to avoid carry-spilling.

Usage of the soft-backend can be forced by setting the clmul_force_soft RUSTFLAG.

$ RUSTFLAGS="--cfg clmul_force_soft" cargo bench

ARMv8 intrinsics (PMULL, nightly-only)

On aarch64 targets including aarch64-apple-darwin (Apple M1) and Linux targets such as aarch64-unknown-linux-gnu and aarch64-unknown-linux-musl, support for using the PMULL instructions in ARMv8’s Cryptography Extensions is available when using the nightly compiler, and can be enabled using the clmul_armv8 RUSTFLAG.

$ RUSTFLAGS="--cfg clmul_armv8" cargo bench

On Linux and macOS, when the clmul_armv8 RUSTFLAG is enabled support for AES intrinsics is autodetected at runtime.

x86/x86_64 intrinsics (CMLMUL)

By default this crate uses runtime detection on i686/x86_64 targets in order to determine if CLMUL is available, and if it is not, it will fallback to using a constant-time software implementation.

For optimal performance, set target-cpu in RUSTFLAGS to sandybridge or newer:

Example:

$ RUSTFLAGS="-Ctarget-cpu=sandybridge" cargo bench

Structs

Clmul

Carryless multiplication